Added models for Deal, Derivative, Invoice, Model
Also added seeders for them too. Haven't done derivative, but not going to be using this at first, so will leave it.

Also adjusted size of integer fields to allow for bigger values in all of the migrations. Particularly for autoincrement fields

+++ app/database/migrations/2014_11_09_152803_create_models_table.php
@@ -16,8 +16,8 @@ class CreateModelsTable extends Migration {
         Schema::create('models', function(Blueprint $table)
         {
             //
-            $table->increments('modelid');
-            $table->tinyInteger('manufacturerid');
+            $table->bigIncrements('modelid');
+            $table->integer('manufacturerid');
             $table->string('modelname')->unique;
             $table->text('description');
             $table->string('website');
